+++ sync_flutter_libs/lib/objectbox_sync_flutter_libs.dart
@@ -18,22 +18,6 @@ Future<Directory> defaultStoreDirectory() async => Directory(
   '${(await getApplicationDocumentsDirectory()).path}/${Store.defaultDirectoryPath}',
 );
 
+// Keep method channel for later use
+// ignore: unused_element
 const _platform = MethodChannel("objectbox_sync_flutter_libs");
-
-/// If your Flutter app runs on Android 6 (or older) devices, call this before
-/// using any ObjectBox APIs, to fix loading the native ObjectBox library.
-///
-/// If the device is running Android 6 (or older) this will try to load the
-/// native library using Java APIs. Afterwards, calling ObjectBox APIs should
-/// load the library successfully on the Dart/Flutter side.
-///
-/// See the [GitHub issue for details](https://github.com/objectbox/objectbox-dart/issues/369).
-Future<void> loadObjectBoxLibraryAndroidCompat() async {
-  if (!Platform.isAndroid) {
-    // To support calling this in multi-platform Flutter apps
-    // do nothing if not Android (plugins for other platforms do not
-    // implement method below).
-    return;
-  }
-  await _platform.invokeMethod<String>('loadObjectBoxLibrary');
-}